refactor: split oversized quota sources
